Split into chunks before writing.
